API Reference

API Reference

Indexes.Tabix

Indexes.overlapchunks

Indexes.Tabix — Method.

Tabix(filename::AbstractString)
Tabix(input::IO)

Load a Tabix index from filename or input.

source

Indexes.overlapchunks — Method.

overlapchunks(tabix::Tabix, interval::Interval)

Return chunks possibly overlapping with the range specified by interval.

Note that records within the returned chunks are not guaranteed to actually overlap the query interval.

source